Extracted the implementation to separate class.
